* 'master' of github.com:shiozaki/BAGEL: (434 commits)
  small fix
  cosmetic
  manual update
  permission
  numerical stability in ci search
  print timing fix
  minor
  implemented multiple force calculation for CASSCF and CASPT2
  caspt2 numerical deriv coup restored
  caspt2grad cleanup
  NacmEval is gone (they are merged to GradEval)
  (3) multiple gradient evaluations work in opt
  (2) intermediate
  intermediate(1)
  manual updates
  Bug fix for serial builds
  Compilation bug fix
  Compilation bug fix
  minor (detected by mpiicpc)
  bug fixed
  ...

# Conflicts:
#	src/molecule/molecule.h

+++ src/mat1e/mixedbasis.h
@@ -62,7 +62,7 @@ class MixedBasis : public MatType {
       }
     }
 
-    void print(const std::string in = "", const int size = 10) const {
+    void print(const std::string in = "", const size_t size = 10) const {
       std::cout << "++++ " << in << " ++++" << std::endl;
       for (int i = 0; i != std::min(size, this->ndim()); ++i) {
         for (int j = 0; j != std::min(size, this->mdim()); ++j) {
